% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanosvgr-package.R
\docType{package}
\name{nanosvgr-package}
\alias{nanosvgr}
\alias{nanosvgr-package}
\title{nanosvgr: Read 'SVG' as Geometric Data}
\description{
Parse SVG to bezier curves, lines and associated graphical parameters. This enables the manipulation of the raw SVG data and custom plotting techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/nanosvgr}
  \item Report bugs at \url{https://github.com/coolbutuseless/nanosvgr/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
