% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanosvg.R
\name{nsvg_add_points}
\alias{nsvg_add_points}
\title{Refresh points data to an nsvg object}
\usage{
nsvg_add_points(nsvg, n = 20)
}
\arguments{
\item{nsvg}{'nsvg' object created using \code{nsvg_read()}}

\item{n}{number of points to use when converting each bezier to a polyline.
Default: 20.}
}
\value{
'nsvg' object with modified 'points' column.  This is a 
        list-column with each element being a data.frame of computed points
        along this path
}
\description{
This is usually done as part of calling \code{nsvg_read()}, but can be
called separately if the points data needs to be calculated at a finer or
coarser level.
}
\examples{
fn <- system.file("sailboat.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) 
# Use a very low resolution conversion. Only 3 points for each bezier.
nsvg <- nsvg_add_points(nsvg, n = 3)

}
