% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nsvg_explode}
\alias{nsvg_explode}
\title{Explode an nsvg object by shifting individual shapes}
\usage{
nsvg_explode(nsvg, scale = 2)
}
\arguments{
\item{nsvg}{'nsvg' object created using \code{nsvg_read()}}

\item{scale}{scale factor for explosion. Default: 2}
}
\value{
Modified 'nsvg' object with all coordinates scaled by the given
        factors
}
\description{
A simple demonstration of how we now have access to all the geometry
}
\examples{
fn <- system.file("sailboat.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) |> nsvg_scale(0.3)
grid::grid.newpage(); plot(nsvg)
nsvg <- nsvg_explode(nsvg, scale = 2)
grid::grid.newpage(); plot(nsvg)

}
