% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nsvg_invert_closedness}
\alias{nsvg_invert_closedness}
\title{Invert the closed-ness of each shape}
\usage{
nsvg_invert_closedness(nsvg)
}
\arguments{
\item{nsvg}{'nsvg' object created using \code{nsvg_read()}}
}
\value{
Modifed 'nsvg' object with the 'closed' status of all paths inverted
}
\description{
Some SVGs will have path closedness incorrectly set.  This is a helper function
to invert the closedness of all paths.  Try this if the plot output is empty.
}
\examples{
fn <- system.file("shuttle.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) |> nsvg_scale(0.3)
grid::grid.newpage(); plot(nsvg)
nsvg <- nsvg_invert_closedness(nsvg)
grid::grid.newpage(); plot(nsvg)

}
