% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanosvg.R
\name{nsvg_read}
\alias{nsvg_read}
\title{Read an SVG as geometry}
\usage{
nsvg_read(filename, n = 20, units = "px", dpi = 96)
}
\arguments{
\item{filename}{SVG filename}

\item{n}{number of points to use when converting each bezier to a polyline.
Default: 20.  Use "NA" to indicate that no conversion should be done.}

\item{units}{units to use. Default 'px'.  One of 'px', 'pt', 'pc' 'mm', 'cm', or 'in'}

\item{dpi}{dots per inch. Default: 96}
}
\value{
An \code{nsvg} object which is a data.frame of SVG data. 
        An SVG is made up of one-or-more \emph{shapes}.
        Each \emph{shape} contains one-or-more \emph{paths}.  Each \emph{path}
        is made of one-or-more \emph{cubic beziers}.  Each shape has a 
        set of graphical parameters
\describe{
  \item{\code{shape_idx}}{[int] Shape index.  Each SVG is defined as a number of shapes, 
        with each shape having a number of \emph{paths}.}
  \item{\code{fill}}{[chr] fill color}
  \item{\code{stroke}}{[chr] stroke color}
  \item{\code{alpha}}{[dbl] opacity in range [0, 1]}
  \item{\code{lwd}}{[dbl] line width}
  \item{\code{linejoin}}{[chr] Line join style.  'bevel', 'mitre' or 'round'}
  \item{\code{lineend}}{[chr] Line end style. 'round', 'butt', 'square'}
  \item{\code{linemitre}}{[dbl] Line mitre limit}
  \item{\code{linedash}}{[list] Raw list of line dash lengths for each shape}
  \item{\code{fill_rule}}{[chr] 'evenodd' or 'winding'}
  \item{\code{fill_type}}{[chr] 'flat', 'linear', 'radial', 'none', 'undef'}
  \item{\code{gradient}}{[list] Radial or linear gradient information for this shape}
  \item{\code{stroke_type}}{[chr] 'flat', 'linear', 'radial', 'none', 'undef'}
  \item{\code{beziers}}{[list] A list of data.frames - one data.frame for each
        shape containing the coordinates of the bezier control points 
        (Note: there are 4 control points for each cubic bezier).
        \describe{
           \item{\code{path_idx}}{[int] Index of path within shape}
           \item{\code{bez_idx}}{[int] Index of bezier with path}
           \item{\code{closed}}{[lgl] Is the path closed?}
           \item{\code{x}}{[dbl] x coordinate of bezier control points}
           \item{\code{y}}{[dbl] y coordinate of bezier control points}
        }
      }
  \item{\code{lty}}{[chr] Line type. Either 'solid' or a string of up to 8 
        characters (from c(1:9, "A":"F")) may be given, giving the length of 
        line segments which are alternatively drawn and skipped.  See
        \code{?graphics::par} for deails on Line Type Specification}
  \item{\code{points}}{[list] A list of data.frames - one data.frame for each
        shape containing the polylines derived from the beziers
        \describe{
           \item{\code{path_idx}}{[int] Index of path within shape}
           \item{\code{bez_idx}}{[int] Index of bezier with path}
           \item{\code{closed}}{[lgl] Is the path closed?}
           \item{\code{x}}{[dbl] x coordinate of polylines}
           \item{\code{y}}{[dbl] y coordinate of polylines}
        }
      }
}
}
\description{
Read an SVG as geometry
}
\examples{
fn <- system.file("sailboat.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) 
head(nsvg[, 1:8])
grid::grid.newpage()
plot(nsvg)
# Manually change the fill color for each shape
nsvg$fill <- topo.colors(nrow(nsvg)) 
grid::grid.newpage()
plot(nsvg)

}
