% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nsvg_scale}
\alias{nsvg_scale}
\title{Scale nsvg object}
\usage{
nsvg_scale(nsvg, scalex = 1, scaley = scalex)
}
\arguments{
\item{nsvg}{'nsvg' object created using \code{nsvg_read()}}

\item{scalex, scaley}{scale factors. Default: 1}
}
\value{
Modified 'nsvg' object with all coordinates scaled by the given
        factors
}
\description{
Scale nsvg object
}
\examples{
fn <- system.file("sailboat.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) 
grid::grid.newpage(); plot(nsvg)

nsvg <- nsvg_scale(nsvg, 0.3)
grid::grid.newpage(); plot(nsvg)

}
