% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nsvg_to_grob}
\alias{nsvg_to_grob}
\title{Conver nsvg object to grid graphics objects}
\usage{
nsvg_to_grob(nsvg, n = 20, inverty = TRUE)
}
\arguments{
\item{nsvg}{nsvg object}

\item{n}{if beziers not yet flatted to points, will use this value}

\item{inverty}{logical. Invert the y-axis? Default: TRUE}
}
\value{
A \code{grid} graphics grob
}
\description{
Conver nsvg object to grid graphics objects
}
\examples{
fn <- system.file("sailboat.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) 
grob <- nsvg_to_grob(nsvg)
grob
grid::grid.draw(grob)

}
