% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanosvg.R
\name{nsvg_unnest_points}
\alias{nsvg_unnest_points}
\title{Unnest SVG polylines into a single long data.frame}
\usage{
nsvg_unnest_points(nsvg, inverty = TRUE, gpars = FALSE)
}
\arguments{
\item{nsvg}{'nsvg' object created using \code{nsvg_read()}}

\item{inverty}{logical. Invert the y-axis? Default: TRUE}

\item{gpars}{logical. Include graphical parameters in result? Default: FALSE}
}
\value{
data.frame with a row for each coordinate for each polyline calculated
        from the original beziers.
}
\description{
Unnest SVG polylines into a single long data.frame
}
\examples{
fn <- system.file("sailboat.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) 
points <- nsvg_unnest_points(nsvg)
head(points)

}
