% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nsvg}
\alias{plot.nsvg}
\title{Plot SVG object}
\usage{
\method{plot}{nsvg}(x, ...)
}
\arguments{
\item{x}{nsvg object}

\item{...}{ignored}
}
\value{
return \code{nsvg} object invisibly
}
\description{
Plot SVG object
}
\examples{
fn <- system.file("sailboat.svg", package = 'nanosvgr', mustWork = TRUE)
nsvg <- nsvg_read(fn) 
plot(nsvg)

}
